+++ src/data/spells/01/anticipate-peril.ts
@@ -1,23 +1,23 @@
 import { Assistant } from "assistant.ts";
 
-export const path = ["Spells", "1st Rank", "Inside Ropes"];
+export const path = ["Spells", "1st Rank", "Anticipate Peril"];
 
 export const actions: Assistant.Action[] = [
     {
         trigger: "spell-cast",
-        predicate: ["item:inside-ropes", { gte: ["item:rank", 3] }],
+        predicate: ["item:anticipate-peril"],
         process: async (data: Assistant.Data) => {
             if (!data.speaker) return;
+            if (!data.target) return;
             if (!data.item?.isOfType("spell")) return;
-            const target = data.target ?? data.speaker;
 
             await game.assistant.socket.addEffect(
-                data.speaker.actor,
-                PF2E_SPELL_EFFECTS["spell-effect-inside-ropes-3rd-rank"],
+                data.target.actor,
+                PF2E_SPELL_EFFECTS["spell-effect-anticipate-peril"],
                 {
                     origin: data.speaker,
                     item: data.item,
-                    target: target
+                    target: data.target
                 }
             );
         }